# Optionally skip json_encoder for orjson