@grigi 2020-05-19
After trying to write tests for this I realise that my assumptions didn't take into account non-complete or non-coerced responses.

Seems that if the response is already a pydantic object it qualifies for the jsonable_encoder to be skipped. (I did tests on this, and it seems to hold)

Looking at the pydantic `ModelField` it's not apparent how to test if the input is of type in the ModelField, of it it is a list/dict.

Seems like we can only do this skipping if validation is turned off for the response?