@grigi 2020-07-26
Hi, I forgot about this sometime during having covid. Thanks for reminding me about this.

@tiangolo I have a few questions:

Assumptions (please correct me if I'm wrong):
* A response can either be a native Pydantic model object(`PMO`), OR a `dict`, OR a manual `Response` type.
* We have no control over a manual `Response` type.
* I'm assuming that `dict` responses will want to be validated and turned into a `PMO` before serialised by default. There is cases where the user wants to say "trust me" and turn off validation, and we should allow this with a global setting.

Questions:
* Can we assume that a native `PMO` is already validated? (You have to jump through serious loops to build an unvalidated `PMO`, and if you do that, one can argue that it's your perogative to ensure that things are correct). If so, can we safely skip the validation on `PMO`'s by default?
* Why did you replace the JSON encoder with your own one, instead of using the provided one from Pydantic: https://pydantic-docs.helpmanual.io/usage/exporting_models/#modeljson ? It already seems to handle all the common data type conversions correctly? Based on quick tests, it appears to be faster too.


Depending on the answer you provide, I can complete this PR. It would be lovely to have a faster response type for the cleaner code by default.

Right now returning a `PMO` (which is the cleanest in code) will result in the slowest response type.
`dict`'s are slightly faster, and `Response` is the fastest.

I would like making `PMO` perform better by default, skip validation and re-building of `PMO`, so we can just go to serialisation directly?

I would like to make the default `JSONResponse` use the provided `PMO.json()` if available so we can honour custom encoders on it.

I would like to make the `ORJSONResponse` allow skipping the datatype conversion by default, as it seems to handle all the common cases as well.

BUT, I would like to limit breakage. So what would be safe behaviour be for `PMO` and `dict` response types?