@raphaelauv 2020-09-23
fastAPI is a API framework doing data validation , so it's the default behavior -> https://github.com/tiangolo/fastapi/issues/1792#issuecomment-665165674

if you want skip validation for performance purposes then it should be explicit The documentation already explain perfectly this :

https://fastapi.tiangolo.com/advanced/response-directly/

`It won't do any data conversion with Pydantic models, it won't convert the contents to any type, etc.`