@kolypto 2020-11-02
It's nice that FastAPI does data validation, but it's best done in development and testing environments only. 
In production, validating APIs that return lists of objects is a significant performance penalty