@lslunis 2021-06-03
@grigi Any update on this? I’d love to be able to disable heavyweight response validation in production.