@grigi 2021-06-04
@lslunis Unfortunately due to not getting answers from FastAPI developers I abandoned the project. I still believe that my findings were valid and the proposal made sense, but its more than a year later, and things might have changed.

Anyone is welcome to take this work and continue/reimplement it.