@ghandic 2021-09-27
@tiangolo what are you thoughts on this? There is the potential for duplicate validation running and a performance hit see https://github.com/tiangolo/fastapi/discussions/3624#discussioncomment-1084046