@MaximilianFranz 2022-06-01
> @tiangolo what are your thoughts on this? There is the potential for duplicate validation running and a performance hit see [#3624 (comment)](https://github.com/tiangolo/fastapi/discussions/3624#discussioncomment-1084046)

Same question for me, this is a very common requirement. Using a custom Response (i.e. `PydanticJSONResponse`) like so: 

```
class PydanticJSONResponse(JSONResponse):
    def render(self, content: Any) -> bytes:
        if content is None:
            return b""
        if isinstance(content, bytes):
            return content
        if isinstance(content, BaseModel):
            return content.json().encode(self.charset)
        return content.encode(self.charset)
```

is a solution, but could also be provided by fastAPI directly.