@tiangolo 2022-11-05
Thanks for the effort @grigi and the conversation everyone!

So, first, a path operation can return anything, not just the types defined above, e.g. an ORM object. The jsonable_encoder in FastAPI will do its best to encode that to something JSON-able. If a Pydantic model is returned, there's no guarantee that it's already validated, the `response_model` can be used to filter data: https://fastapi.tiangolo.com/tutorial/response-model/#add-an-output-model

I want to add a way to allow users to customize how data is serialized and provide their own functions for serialization and deserialization. That would probably help here, as people would be able to use a function that doesn't use `jsonable_encoder` and just uses Pydantic's own mechanisms.

Also, with the soon-to-be-released Pydantic v2, a lot of the duplication cloning Pydantic fields in `response_model` and similar things will be avoided.

> Sorry for the long delay! 🙈  I wanted to personally address each issue/PR and they piled up through time, but now I'm checking each one in order.